Apply routes to external vehicles #skip_bins
Previously route assignements to externally controlled entities were ignored since its position is not updated by the scenario anyway. But if a vehicle is bound to a route it should affect the mapping of XYZ position to the closest position on the road network (XYZH2TrackPos function), in the sense that it searches closest position on the route. In effect limiting the search scope to only those roads present in the route waypoints. So, routes are now also assigned to externally controlled vehicles, even if they do not affect the actual position, only calculated road coordinates.