Support any synch point #skip_bins
As pointed out by issue 48, the synchronizeAction currently does not
work when target position for action object or target object is not
along a route on the road network. This is a unnecessary limitation.
By this commit the calculation will fall back to inertia (X, Y) distance
when road network path to target can't be identified to target.