Fix issue in intersection path planning #skip_bins
When mapping a xyz position to road coordinates, the search terminated
as soon as a road was found that contained the point. However, in
junctions there might be overlapping roads and hence the point might
match multiple roads. Therefor the updated strategy is to search all
roads and choose the best one (given a few criterias).